[INPLACE-249] 게시글 작성 로직 구조 분리
